Replace deep_merge by recursive minimum_check
